backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=1.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0
numpy>=1.24.0
